backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
pydantic[email]>=2.9.0
pydantic-settings>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
alembic>=1.13.0

# Logging
structlog>=24.4.0
python-json-logger>=2.0.7

# Rate limiting
slowapi>=0.1.9

# WebSocket
websockets>=12.0

# Caching
redis>=5.0.0
hiredis>=2.3.0

# Export
openpyxl>=3.1.0
pandas>=2.2.0

# Email
aiosmtplib>=3.0.0
blinker>=1.7.0

# 2FA
pyotp>=2.9.0
qrcode[pil]>=7.4.2

# File handling
python-multipart>=0.0.12
aiofiles>=23.2.0
Pillow>=10.0.0
python-magic>=0.4.27

# SSO / SAML
python3-saml>=1.16.0

# AI / ML
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.2.0

# GraphQL
strawberry-graphql>=0.210.0
strawberry-graphql[fastapi]>=0.210.0

# System monitoring
psutil>=5.9.0
sentry-sdk[fastapi]>=1.40.0
prometheus-client>=0.19.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
httpx>=0.27.0
anyio>=4.4.0
ruff>=0.6.0
black>=24.8.0
mypy>=1.11.0
